requirements.txt
django==1.11
djangorestframework
django-cors-headers
djangorestframework-jwt
